fix defaultKeyBindingPlugin usage & move editor utils to the editor
