+++ draft-js-plugins-editor/src/Editor/createCompositeDecorator.js
