+++ draft-js-plugins-editor/src/Editor/moveSelectionToEnd.js
